requirements.txt
selenium
webdriver-manager
pymysql
sqlalchemy
python-dotenv
pillow